refactor: 보드와 기물 생성 책임을 factory로 분리
